refactor: rename Options to CardOptionsStore
